perf: allocate distributional sample grid as float32 (#46)
sample_prediction_grid allocated the (N, n_samples) sample table as
float64 and held all targets live at once (3 x ~2.4 GB on global-land
pgm), peaking ~12 GB and OOM-killing constituent runs on 31 GB hardware.

The output PredictionFrame is already float32 (coerce-enforced, ADR-020),
so the float64 buffer was downcast-and-discarded at the to_prediction_frames
seam. Allocating float32 up front just moves that existing downcast earlier:
output is byte-identical (golden tests pass with zero regeneration) at half
the grid memory. Draw order (ADR-011 §3) is untouched; history pools stay
float64 (test_pooling byte-identity, C-32).

This is Option A of #46 (the free, order-safe half). Per-target streaming
(the ~3x-further, RNG-reordering half) is tracked separately as a deliberate
re-baseline.

231 passed, 1 xfailed; ruff clean.

+++ views_baseline/model/frames/output.py
@@ -128,7 +128,12 @@ def sample_prediction_grid(
     n_rows = len(time_arr)
     rng = np.random.default_rng(seed)
 
-    y_preds = {t: np.empty((n_rows, n_samples), dtype=np.float64) for t in targets}
+    # float32 matches the PredictionFrame's coerce-enforced dtype (ADR-020): the float64
+    # buffer was downcast-and-discarded at the to_prediction_frames seam, so allocating
+    # float32 up front is byte-identical output at half the memory — it halves the
+    # distributional grid footprint (issue #46). Draw order (ADR-011 §3) and the golden
+    # values are unchanged. History pools stay float64 (test_pooling byte-identity, C-32).
+    y_preds = {t: np.empty((n_rows, n_samples), dtype=np.float32) for t in targets}
     idx = 0
     for cid in entities:
         for _ in time_ids:
